requirements.txt

autopep8==1.5.4
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
cryptography==3.3.1
cycler==0.10.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2
django-cleanup==5.1.0
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-storages==1.11.1
filelock==3.0.12
google==3.0.0
gunicorn==20.0.4
idna==2.10
isort==5.6.4
joblib==1.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
pipenv==2020.11.15
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sortedcontainers==2.3.0
soupsieve==2.2.1
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.2
urllib3==1.26.2
virtualenv==20.2.2
virtualenv-clone==0.5.4
whitenoise==5.2.0
wrapt==1.12.1
